+++ internal/tenant/test_helpers_test.go
@@ -29,7 +29,6 @@ func newTestDatabaseWithLogger(t *testing.T, customLogger logger.Interface) *gor
 	if err := dbInstance.AutoMigrate(
 		&Tenant{},
 		&TenantDomain{},
-		&SenderDomain{},
 		&EmailProfile{},
 		&SMSProfile{},
 	); err != nil {